+++ projects/ng-virtual-list/CHANGELOG.md
@@ -1,6 +1,11 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 
+## [20.11.11] - 2026-04-19
+
+### Fixed
+- Fixed a freeze when deleting a list.
+
 ## [20.11.10] - 2026-04-13
 
 ### Fixed